requirements.txt
requests
pyotp==2.9.0
https://github.com/Shoonya-Dev/ShoonyaApi-py/raw/refs/heads/master/dist/NorenRestApi-0.0.30-py2.py3-none-any.whl
pandas==2.2.3
zipfile2
PyYAML==6.0.2
selenium==4.27.1
# upstox-python-sdk
upstox-python-sdk==2.9.0
# ta==0.11.0
# yfinance==0.2.54
